% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ckan.R
\name{search_ckan}
\alias{search_ckan}
\alias{search_ckan_vic}
\alias{search_ckan_nsw}
\alias{search_ckan_aus}
\alias{search_ckan_qld}
\alias{search_ckan_sa}
\title{Search ckan data repository}
\usage{
search_ckan(search_term, ckan_url, rows = 10, detailed = FALSE)

search_ckan_vic(search_term, ...)

search_ckan_nsw(search_term, ...)

search_ckan_aus(search_term, ...)

search_ckan_qld(search_term, ...)

search_ckan_sa(search_term, ...)
}
\arguments{
\item{ckan_url}{}
}
\value{

}
\description{
Search ckan data repository
}
